New code to handle stability figures for sub_to_cor
Added new figures to assess the "stability" of the results from the
subcortical to cortical correlation figures based on reviewer feedback.
Specifically, one new function returns the 3D coordinates of the top and
bottom N voxels for the contrast or body pRF analyses; the next function
takes these coordinates and computes the result of both methods of the
subcortical to cortical correlation analysis.

Subsequent commits will take those outputs and add code to handle the
visualization and plotting of the resulting correlation profiles.